--- presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/uistate/SeasonUiState.kt
@@ -1,16 +1,0 @@
-package com.berlin.aflami.viewmodel.uistate
-
-data class EpisodeUi(
-    val id: Int,
-    val episodeNumber: Int,
-    val title: String,
-    val description: String,
-    val imageUrl: String,
-    val time: String,
-    val date: String,
-    val rating: String,
-)
-data class SeasonUi(
-    val seasonNumber: String,
-    val episodes: List<EpisodeUi>
-)
